refactor: ♻️ interaction~ variant 적용
